hltrailingws: adjust keepautoindent=off implementation
Fix undesirable highlighting of whitespace in empty line when we
insert 2 newlines and then undo the 2nd newline.
To fix it, change the order of operations. Text operations are undone
in reverse order, so autoremove of whitespace in previous line
should be done first, so that when InsertNewline is undone, inserting
this autoremoved whitespace is done last, so that the cursor doesn't
move between lines after that.